feat: max concurrency per environment for function executions
We currently only have a global concurrency limit at the processor level
that limit how many functions are executed at the same time. It is nice
to protect our infra but if a large number of tasks is queued for the
same environment they are all dispatched to a single runner that cannot
handle the load.
The orchestrator has a mechanism to limit how many tasks for a specific
group can be executed at the same time. We currently only use to to
limit the concurrency of async actions per environment.
I propose to set a upper-limit of how many functions can be executed per
environment and type of functions to avoid dispatching tons of functions
to a single runner at the same time.
This commit introduce global limit per environment and per function
types but a possible improvement would be to have dedicated values in
plan (the same way we are doing for the API rate-limit) so different
customers can have different limits