fix: reduce requests to Slack
Slack notification: we currently join the channel every time a Slack
notification is sent, even if it is not necessary.
This commit modifies the flow, first trying the send the message and
only joining if we get a `not_in_channel` error. That's an extra call
for the first notification but all the following one will be one request
instead of two.