fix(otlp-exporter): include /v1 prefix automitically
For some reason I don't recall I didn't include /v1 in the
OTLPTraceExporter url path and require the customer to add it in the
otlp collector settings.
This was an error, most of the customers are confused by it even though
it is in the docs and in the placeholder example.
In order not to break existing customers, we are making the /v1 optional
with this commit. Once I have fixed all customers settings I will remove
the call to replace()