feat(billing): allow billing batcher to group/aggregate similar events
It would be mostly useful for billing events like proxy, function_executions of
actions and webhooks but I have implemented in a generic way so it works
for all events (now and future)

It would have been quite hard to aggregate billing events in a type-safe
way because most of the properties were Record<string, ...>
So I started with a refactor of the billing types and make each
attributes explicitly defined. The drawback is that we cannot add a new
field and have it automatically send to orb without modifying the
billing package but without it we cannot have safe grouping. It might be
a bit annoying at the moment where we are adding/tweaking billing events
a lot but I imagine it would be pretty stable at some point.

Then, since some of the properties of billing events are dimensions and
some are values, we must be able to defined which attributes to group
against and which attributes to aggregate/sum. I first started with a
fully type-safe implementation at compile time. It worked but it was 50
lines of typescript types magic that I was not very confortable with so I went
a simpler route that relies on tests.
For every type of billing event, we must specify which attributes are
being used for defining a unique key used for grouping. It is done by
including attributes by default and forces us to explicitely omit them
if they are not a grouping attributes. Like this, we cannot forget since
the grouping key would change and fail the tests.