fix(orch): simplify dequeue query
with the introduction of multiple sync groups, the dequeue query is
randomly blocking when calculating the count of running tasks for each
group.
This commit simplify the `running` CTE query, not making it dependent on
the `candidates` CTE and calculating running count for all sync groups.
It will result in some groups not be used because they don't have an
candidates tasks but the CTE will be consistently fast (using the
idx_tasks_group_key_started index) wthout a more expensive `IN subquery`