fix: add frequencyMs to records billing events
In order to simplify the calculation of metrics like records or
connections, we are sending the interval at which the data is being
ingested. In orb we then attributes the value to a 1/Nth of the month
(N being the interval). Ex: we send records count every hour = each value
in Orb count for 1h in the month that are summed to arrive at the total
cost
Limitations of this approach:
- granularity is linked to our ingestion frequency.
- "missed" events are not recoverable and therefore not billable
The main advantage: it simplifies both the ingestion code and the Orb
metrics calculation