fix(ci): do not push image and publish package from merge queue
We use the commit hash for package version and image tag.
Since merging queue actually runs on the same commit that is then added
to master the package publish and image push fail after the merge to
master, trying to push/publish the same existing version.
This commit modify the gh workflow so publishing the package and pushing
the image doesn't happen on merge queue