fix: otlp export of actions
logContext operation can be created/ended across services.
Action operations for example are created in server but can be marked as
successful/failed in jobs. In order for the attached otlp span to be
correctly ended, we need to reattach it.
This is an known limit of our otlp export feature where span don't
persist accross services
I haven't look to deep into the git history but I think action
operations used to be created and ended in server but with the
introduction of async actions, the operations must now be ended in jobs
since it is not inline anymore.